Refactor Trajectory class to remove commented code and improve code readability
