+++ fog_x/loader/__init__.py
@@ -1,3 +1,4 @@
 from .base import BaseLoader
 from .rlds import RLDSLoader
 from .hdf5 import HDF5Loader
+from .vla import VLALoader
